Vessel Decoration Operations

+++ IsolatedIslandGame/IsolatedIslandGame.Library/CommunicationInfrastructure/Operations/Handlers/UserOperationHandlers/LoginHandler.cs
@@ -30,7 +30,7 @@ namespace IsolatedIslandGame.Library.CommunicationInfrastructure.Operations.Hand
                         { (byte)LoginResponseParameterCode.FacebookID, player.FacebookID.ToString() },
                         { (byte)LoginResponseParameterCode.Nickname, player.Nickname },
                         { (byte)LoginResponseParameterCode.Signature, player.Signature },
-                        { (byte)LoginResponseParameterCode.GroupType, player.GroupType },
+                        { (byte)LoginResponseParameterCode.GroupType, (byte)player.GroupType },
                         { (byte)LoginResponseParameterCode.LastConnectedIPAddress, player.LastConnectedIPAddress.ToString() },
                     };
                     SendResponse(operationCode, responseParameters);
